refactored sass implementation
vite already supports sass hot-reloading so all the parralels stuff and
the watcher aren't needed.
The minified sass-build step also caused issues because it minified the
css files in src. (instead of dist)
Vite can also build and minify the scss files so we can leave those
commands out to avoid issues.

This means we no longer need the "css" files at all in our src, only the
"scss" files.
Because it is now just one folder, I've opted to put scss in src instead
of nesting it in assets. I also had a short conversation about this with
david and he seemed to really prefer css/scss files to be direclty in
src.

Lastly I replaced "sass" with "sass-embeded" which is by the same
authors/maintainers but more performant.

The scss folder structure is as followes
-scss
  - abstracts  // mixins or other reusable functions
  - components // scss module files for the components
  - pages      // scss module files for the pages
  - partials   // parts of files that are meant to be included in others
  - main.scss  // main file in root of scss folder, global config.